+++ src/workflow-targets/index.js
@@ -1,5 +1,6 @@
 module.exports = {
   custom: require("./custom"),
+  dita: require("./dita"),
   fs: require("./fs"),
   word: require("./word"),
   dsoToepassingsprofielenRhino: require("./dso-toepassingsprofielen-rhino"),